- 멤버십 업로딩 데이터 단위 변경

+++ mago3d-admin/src/main/java/gaia3d/domain/membership/Membership.java
@@ -22,7 +22,7 @@ public class Membership extends Search implements Serializable {
 	// 멤버십 이름
 	private String membershipName;
 	// 허용 용량
-	private Double uploadFileSize;
+	private Integer uploadFileSize;
 	// 허용 횟수
 	private Integer convertFileCount;
 }
